@@ -1,6 +1,6 @@
 {
   "name": "@doerjs/cli",
-  "version": "0.0.2-alpha.8",
+  "version": "0.0.2-alpha.9",
   "description": "",
   "author": "FaureWu <502556093@qq.com>",
   "license": "MIT",
@@ -52,9 +52,9 @@
     "postcss-normalize": "^10.0.1",
     "postcss-preset-env": "^7.4.1",
     "prettier": "^2.5.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^6.2.1",
+    "react": "18.1.0",
+    "react-dom": "18.1.0",
+    "react-router-dom": "6.3.0",
     "resolve": "^1.22.0",
     "semver": "^7.3.5",
     "source-map-loader": "^3.0.1",
