@@ -1,6 +1,6 @@
 module.exports = `
 import React from 'react'
-import ReactDOM from 'react-dom'
+import ReactDOM from 'react-dom/client'
 
 import App from './App'
 
@@ -8,10 +8,8 @@ import { onRender } from './global'
 import { isFunction } from './helper'
 
 function render(Node) {
-  ReactDOM.render(
-    Node,
-    document.getElementById(process.env.ROOT_ELEMENT_ID || 'root'),
-  )
+  const root = ReactDOM.createRoot(document.getElementById(process.env.ROOT_ELEMENT_ID || 'root'));
+  root.render(Node);
 }
 
 if (isFunction(onRender)) {
